refactor: Update date_start and date_end types to Date (#292)
This commit updates the types of the `date_start` and `date_end` properties in the
`CareerTimelineItemProps` and `CareerTimelineProps` interfaces to be of type `Date` instead of
`string`. This change ensures better type safety and consistency when working with dates throughout
the codebase. It also includes the formatting of dates to a more user-friendly format in the `CareerTimelineItem` component.

---------

Co-authored-by: Eric Büttner <eric.buettner@tuffz.com>