feat(utils-locations)!: improve structureLocation function, set `state` to `null` if same as `city`
Refactor the `structureLocation` function to make it more concise and maintainable. Instead of
multiple if statements, use object property shorthand and a ternary operator to set the `state`
property of the `StructuredLocation` object. If `state` is the same as `city`, it is explicitly set
to `null` to avoid redundancy.

BREAKING CHANGE: The behavior of the `structureLocation` function has changed, which may impact code
relying on the previous behavior. This change is considered breaking because it alters the way the
function handles the `state` property when it's the same as `city`.