refactor(career-timeline)!: remove obsolete Month enum
This commit removes the now unnecessary Month enum, previously used for date initialization in the
CareerTimeline component. While the primary motivation is to enhance code clarity and
maintainability, it's worth noting that this change falls under a refactoring category, as it
doesn't alter the external behavior of the CareerTimeline component.

Change:
- Deleted the obsolete Month enum.

BREAKING CHANGE: This removal should not impact the functionality of the CareerTimeline component
itself, and it is considered a refactor rather than a breaking change.