ci(renovate): update Renovate package rule for Nx updates
This commit modifies the Renovate package rule related to Nx updates in the `renovate.json`
configuration file. Previously, the rule included a post-upgrade task intended for running
migrations on Nx updates. However, this task was found to be buggy or not feasible for Renovate.
Consequently, the post-upgrade task has been removed from the rule.

Additionally, the automerge property is retained and set to false to prevent automerging for Nx
updates. This ensures that changes related to Nx updates will not be automatically merged.

ref #475