refactor(coding-booth-com-feature-character-counter)!: rename library
This commit refactors the `character-counter` library within the `coding-booth-com` scope by
renaming it to `feature-character-counter`.

Included changes:

- Renaming the `README.md` file from `character-counter` to
`coding-booth-com-feature-character-counter`.
- Renaming the `project.json` file from `character-counter` to
`coding-booth-com-feature-character-counter`.
- Renaming various files and directories associated with the `character-counter` library to
`coding-booth-com-feature-character-counter`.
- Updating references in the tsconfig.base.json file from
`@tuffz/coding-booth-com/character-counter` to `@tuffz/coding-booth-com-feature-character-counter`.

This refactor introduces a breaking change as it involves renaming key files and directories.
However, the `character-counter` was renamed to `coding-booth-com-feature-character-counter` to
better align with the project's structure and naming conventions.