build(core): adjust semantic commit type to 'fix' in the 'monorepo:nextjs' preset (Renovate)
Update the description and remove unnecessary settings for adjusting the semantic commit type to
'fix' in the 'monorepo:nextjs' preset.

It's recommended to keep the group setting and set the commit type to 'fix' regardless of whether
it's a devDependency or a dependency.

More details can be found in the discussion:
https://github.com/renovatebot/renovate/discussions/8875