build(core): adjust semantic commit type to 'fix' in the 'monorepo:nextjs' preset (Renovate) (#546)
* build(core): adjust semantic commit type to 'fix' in the 'monorepo:nextjs' preset (Renovate)

Update the description and remove unnecessary settings for adjusting the semantic commit type to
'fix' in the 'monorepo:nextjs' preset.

It's recommended to keep the group setting and set the commit type to 'fix' regardless of whether
it's a devDependency or a dependency.

More details can be found in the discussion:
https://github.com/renovatebot/renovate/discussions/8875

* ci(github-actions): add Renovate config validation job

Added a new job named "Renovate Config Validation" to validate the Renovate configuration file. This
Job ensures the validity of the Renovate configuration with strict mode enabled.